requirements.txt
asgiref==3.11.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.15.1
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0